Configure deployment, update README, and fix bugs

+++ flask1/users/utils.py
@@ -20,7 +20,7 @@ def save_picture(form_picture):
     i.thumbnail(outpu_size)
 
     i.save(picture_path)
-    if current_user.image_file:
+    if current_user.image_file and current_user.image_file != 'default.jpg':
         old_picture_path=os.path.join(current_app.root_path, "static/profile_pics", current_user.image_file)
         if os.path.exists(old_picture_path):
             os.remove(old_picture_path)
@@ -29,17 +29,32 @@ def save_picture(form_picture):
 
 
 def send_reset_email(user):
-    token=user.get_reset_token()
-    msg = MIMEText(f"""To reset your password click the following link:
+    try:
+        token=user.get_reset_token()
+        msg = MIMEText(f"""To reset your password click the following link:
 {url_for("users.reset_password", token=token, _external=True)}
 If you did not make this request then simply ignore this email and no changes will be made
 """)
-    msg["Subject"] = "Password Reset Request"
-    msg["From"] = current_app.config["MAIL_USERNAME"]
-    msg["To"] = f"{user.email}"
+        msg["Subject"] = "Password Reset Request"
+        
+        mail_user = current_app.config.get("MAIL_USERNAME")
+        mail_pass = current_app.config.get("MAIL_PASSWORD")
+        mail_server = current_app.config.get("MAIL_SERVER")
+        mail_port = current_app.config.get("MAIL_PORT")
+        
+        if not mail_user or not mail_pass:
+            print("SMTP Credentials not set.")
+            return False
 
-    server = smtplib.SMTP(current_app.config["MAIL_SERVER"], current_app.config["MAIL_PORT"])
-    server.starttls()
-    server.login(current_app.config["MAIL_USERNAME"], current_app.config["MAIL_PASSWORD"])
-    server.sendmail(current_app.config["MAIL_USERNAME"], f"{user.email}", msg.as_string())
-    server.quit()
+        msg["From"] = mail_user
+        msg["To"] = f"{user.email}"
+
+        server = smtplib.SMTP(mail_server, mail_port)
+        server.starttls()
+        server.login(mail_user, mail_pass)
+        server.sendmail(mail_user, f"{user.email}", msg.as_string())
+        server.quit()
+        return True
+    except Exception as e:
+        print(f"SMTP Error: {e}")
+        return False
